小更新

+++ MaxTunnel/MaxTunnel-Web/src/main/webapp/WEB-INF/src/views/UM/EM/RelatedUnits/QueryRelatedUnits.vue
@@ -3,60 +3,34 @@
         <Row class="queryCondition">
             <Col span="6">
                 <span>所属管仓段：</span>
-                <!-- <Dropdown style="margin-left: -60px;">
-                    <a href="javascript:void(0)">
-                        <Input type="text" icon="arrow-down-b" style="width:100%" :value="sectionName" :style="{width:w+'px',marginLeft:'60px'}"></Input>
-                    </a>
-                    <DropdownMenu slot="list" v-for="(item,index) in treeList" :key="index">
-                        <Dropdown placement="right-start">
-                            <DropdownItem>
-                                {{item.name}}
-                                <Icon type="ios-arrow-right"></Icon>
-                            </DropdownItem>
-                            <DropdownMenu slot="list">
-                                <DropdownItem v-for="(firstChild,index) in item.list" :key="index"> 
-                                    <Dropdown placement="right-start" v-if="firstChild.list && firstChild.list.length >0">
-                                        <DropdownItem style="padding: 0px">
-                                            {{firstChild.name}}
-                                            <Icon type="ios-arrow-right"></Icon>
-                                        </DropdownItem>
-                                        <DropdownMenu slot="list">
-                                            <DropdownItem v-for="(secondChild,index) in firstChild.list" :key="index" v-model="conditions.sectionIds" @click.native='doMore(secondChild.name,secondChild.id)'>{{secondChild.name}}</DropdownItem>
-                                        </DropdownMenu>
-                                    </Dropdown> 
-                                    <DropdownItem v-else style="padding: 0px;">
-                                        {{firstChild.name}}    
-                                    </DropdownItem>    
-                                </DropdownItem>
-                            </DropdownMenu>
-                        </Dropdown>
-                    </DropdownMenu>
-                </Dropdown>  -->
                 <Input type="text" icon="arrow-down-b" style="width: 60%" :value="sectionName" @click="showTree"></Input>                   
                 <Tree class="tree"  :data="treeList" show-checkbox ref="tree" @on-check-change="choiceAll" v-show="isShow"></Tree>
             </Col>
             <Col span="6">
                 <span>单位名称：</span>
-                <Input type="text" v-model="conditions.name" id="getW"  style="width:60%"  @on-keyup="queryUnits()"></Input>
+                <Input type="text" v-model="conditions.name" id="getW"  style="width:60%"></Input>
             </Col>
             <Col span="6">
                 <span>联系人：</span>
-                <Input type="text" v-model="conditions.contact"  style="width:60%"  @on-keyup="queryUnits()"></Input>
+                <Input type="text" v-model="conditions.contact"  style="width:60%"></Input>
             </Col>
             <Col span="6">
                 <span>单位类别：</span>
-                <Select v-model="conditions.unitType"  style="width:60%" @on-change="queryUnits()">
+                <Select v-model="conditions.unitType"  style="width:60%">
                     <Option value=null>所有</Option>
                     <Option v-for="(item,index) in unitType" :value="item.val" :key="index">{{ item.key }}</Option>
                 </Select>
             </Col>
             <Col span="6">
                 <span class="word025">开始时间：</span>
-                <DatePicker type="datetime" placeholder="请选择开始时间"  @on-change="queryUnits()" style="width: 60%" v-model="conditions.startTime"></DatePicker>
+                <DatePicker type="datetime" placeholder="请选择开始时间"  style="width: 60%" v-model="conditions.startTime"></DatePicker>
             </Col>
             <Col span="6">
                 <span>结束时间：</span>
-                <DatePicker type="datetime" placeholder="请选择结束时间"  @on-change="queryUnits()" style="width: 60%" v-model="conditions.endTime"></DatePicker>
+                <DatePicker type="datetime" placeholder="请选择结束时间"  style="width: 60%" v-model="conditions.endTime"></DatePicker>
+            </Col>
+            <Col span="6" offset="6">
+                <Button type="primary" icon="ios-search" @click="queryUnits()">查询</Button>
             </Col>
         </Row> 
         <div class="list">   
@@ -103,11 +77,14 @@
 </template>
 <script>
 import axios from "axios";
-import { TunnelService } from '../../../../services/tunnels'
-import { EnumsService } from '../../../../services/enums'
-import { RelatedUnitService } from '../../../../services/relatedUnits'
+import { TunnelService } from '../../../../services/tunnelService'
+import { EnumsService } from '../../../../services/enumsService'
+import { RelatedUnitService } from '../../../../services/relatedUnitService'
 import Enum from "../../../../../static/Enum.json";
 // import oneTree from '../../../../components/UM/EM/oneTree'
+//瀑布流
+// import Waterfall from '../../../../../node_modules/vue-waterfall/lib/waterfall'
+// import WaterfallSlot from '../../../../../node_modules/vue-waterfall/lib/waterfall-slot'
 export default {
     // components:{
     //     oneTree
@@ -227,47 +204,8 @@ export default {
             function(error){
                 console.log(error)
         })
-        // axios.get('tunnels/tree').then(response=>{
-        //     let{ code,data } = response.data
-        //     if(code==200){
-        //         data.forEach(element => {
-        //             var firstChildren = []
-        //             let temp={}
-        //             temp.title = element.name
-        //             temp.id = element.id
-        //             if(element.list.length!=0){
-        //                 element.list.forEach(a=>{
-        //                     var secondChildren = []
-        //                     let obj = {}
-        //                     obj.title = a.name
-        //                     obj.id = a.id
-        //                     obj.disabled = true
-        //                     if(a.list.length!=0){
-        //                         obj.disabled = false
-        //                          a.list.forEach(b=>{
-        //                             let secondInfo = {}
-        //                             secondInfo.title = b.name
-        //                             secondInfo.id = b.id
-        //                             secondChildren.push(secondInfo)
-        //                             obj.children = secondChildren
-        //                         })
-        //                     }
-        //                     firstChildren.push(obj)                        
-        //                     temp.children = firstChildren
-        //                 })
-        //             }
-        //             this.treeList.push(temp)                   
-        //         });
-        //     }
-        // })
-        // axios.get('/unittype-enums').then(response=>{
-        //     let{ code,data } = response.data
-        //     if(code==200){
-        //         this.unitType = data;
-        //     }
-        // })
         this.queryUnits()
-    },
+    },  
     methods:{
         queryUnits(){
             let _this = this
@@ -290,40 +228,25 @@ export default {
                 },function(error){
                     console.log(error)
             })
-            // axios.post('relatedunits/datagrid',this.params).then(response=>{
-            //     let { code, data } = response.data;
-            //     if (code == 200) {
-            //         this.page.pageTotal = data.total;
-            //         this.unitInfo = data.list;
-            //         for(let index in this.unitInfo){
-            //             this.unitInfo[index].crtTime = new Date(this.unitInfo[index].crtTime).format('yyyy-MM-dd hh:mm:s')
-            //             var str = ''
-            //             for (let k in this.unitInfo[index].sectionList){
-            //                 if(this.unitInfo[index].sectionList.length==1){
-            //                     this.unitInfo[index].sectionIds=this.unitInfo[index].sectionList[k].name                               
-            //                 }else{
-            //                     str += this.unitInfo[index].sectionList[k].name+','
-            //                     this.unitInfo[index].sectionIds = str.substring(0,str.length-1)
-            //                 }
-            //             }
-            //         }
-            //     }
-            // })
         },
         del(index){
-            let _this = this
-            RelatedUnitService.deleteRelatedUnit(_this.unitInfo[index].id).then(
-                function(result){
-                    _this.unitInfo.splice(_this.unitInfo[index].id, 1);
-                    _this.queryUnits();
+            this.$Modal.confirm({
+                title: '相关单位',
+                content: '<p>是否删除这条相关单位信息</p>',
+                onOk: () => {
+                    let _this = this
+                    RelatedUnitService.deleteRelatedUnit(_this.unitInfo[index].id).then(
+                        function(result){
+                            _this.unitInfo.splice(_this.unitInfo[index].id, 1);
+                            _this.queryUnits();
+                        },
+                        function(error){
+                            console.log(error)
+                        }
+                    )
                 },
-                function(error){
-                    console.log(error)
-                })
-            // axios.delete('relatedunits/'+this.unitInfo[index].id).then(response=>{
-            //     this.unitInfo.splice(this.unitInfo[index].id, 1);
-            //     this.queryUnits();
-            // })
+                onCancel: () => {}
+            })
         },
         goToMoudle: function(index, type) {
             this.$router.push({
